Make the class a Mongo Document instead of a JPA entity

+++ src/main/java/com/tikalk/workshop/entity/Person.java
@@ -3,6 +3,7 @@ package com.tikalk.workshop.entity;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigInteger;
 import java.util.Objects;
 
 /**
@@ -11,8 +12,9 @@ import java.util.Objects;
  */
 @Document(collection = "persons")
 public class Person {
+
     @Id
-    private Long id;
+    private BigInteger id;
 
     private String givenName;
 
@@ -26,7 +28,7 @@ public class Person {
         this.sureName = sureName;
     }
 
-    public Long getId() {
+    public BigInteger getId() {
         return id;
     }
